Separate selectors for siloDomainData
- Use match for selectors in siloDomainData
