+++ frontend/src/common/selectors/domainData.js
@@ -5,97 +5,155 @@ import { activeUserSelector } from './auth';
 const emptyList = [];
 const emptyObject = {};
 
-
 // Using props
 
 export const userIdFromRoute = (state, { match }) => match.params.userId;
 export const groupIdFromRoute = (state, { match }) => match.params.userGroupId;
+export const countryIdFromProps = (state, { match }) => match.params.countryId;
+export const projectIdFromRoute = (state, { match }) => match.params.projectId;
+
+// Used in silo-reducers
 export const analysisFrameworkIdFromProps = (state, { match }) => match.params.analysisFrameworkId;
 export const categoryEditorIdFromProps = (state, { match }) => match.params.categoryEditorId;
+export const leadIdFromRoute = (state, { match }) => match.params.leadId;
+export const categoryEditorIdFromRoute = (state, { match }) => match.params.categoryEditorId;
 
 export const regionIdFromProps = (state, { regionId }) => regionId;
 export const userGroupIdFromProps = (state, { userGroupId }) => userGroupId;
 export const analysisFrameworkIdFromPropsForProject = (state, { afId }) => afId;
 export const categoryEditorIdFromPropsForProject = (state, { ceId }) => ceId;
 
-// Using state
+// COMMON
 
 export const leadFilterOptionsSelector = ({ domainData }) => (
     domainData.leadFilterOptions || emptyObject
 );
 
-export const entriesSelector = ({ domainData }) => (
-    domainData.entries || emptyObject
+export const visualizationSelector = ({ domainData }) => (
+    domainData.visualization || emptyObject
 );
 
 export const regionsSelector = ({ domainData }) => (
     domainData.regions || emptyObject
 );
 
+export const projectsSelector = ({ domainData }) => (
+    domainData.projects || emptyObject
+);
+
+export const projectsOptionsSelector = ({ domainData }) => (
+    domainData.projectsOptions || emptyObject
+);
+
 export const analysisFrameworksSelector = ({ domainData }) => (
     domainData.analysisFrameworks || emptyObject
 );
 
-export const categoryEditorsSelector = ({ domainData }) => (
-    domainData.categoryEditors || emptyObject
-);
-
-export const analysisFrameworkDetailSelector = createSelector(
-    analysisFrameworksSelector,
-    analysisFrameworkIdFromPropsForProject,
-    (analysisFrameworks, afId) => (
-        analysisFrameworks[afId] || emptyObject
-    ),
-);
-
-export const categoryEditorDetailSelector = createSelector(
-    categoryEditorsSelector,
-    categoryEditorIdFromPropsForProject,
-    (categoryEditors, ceId) => (
-        categoryEditors[ceId] || emptyObject
-    ),
-);
-
-export const analysisFrameworkListSelector = createSelector(
-    analysisFrameworksSelector,
-    analysisFrameworks => (
-        analysisFrameworks && Object.values(analysisFrameworks).filter(
-            analysisFramework => analysisFramework,
-        )) || emptyList,
-);
-
-export const categoryEditorListSelector = createSelector(
-    categoryEditorsSelector,
-    categoryEditors => (
-        categoryEditors && Object.values(categoryEditors).filter(
-            categoryEditor => categoryEditor,
-        )) || emptyList,
-);
-
-export const countriesListSelector = createSelector(
-    regionsSelector,
-    regions => (
-        regions && Object.values(regions).filter(region => region && region.public)
-    ) || emptyList,
-);
-
 export const adminLevelsSelector = ({ domainData }) => (
     domainData.adminLevels || emptyObject
-);
-export const usersSelector = ({ domainData }) => (
-    domainData.users || emptyObject
-);
-export const projectsSelector = ({ domainData }) => (
-    domainData.projects || emptyObject
-);
-export const projectsOptionsSelector = ({ domainData }) => (
-    domainData.projectsOptions || emptyObject
 );
 
 export const groupsSelector = ({ domainData }) => (
     domainData.userGroups || emptyObject
 );
 
+const usersSelector = ({ domainData }) => (
+    domainData.users || emptyObject
+);
+
+const categoryEditorsSelector = ({ domainData }) => (
+    domainData.categoryEditors || emptyObject
+);
+
+// CONSTANTS
+
+export const hierarchialDataSelector = createSelector(
+    visualizationSelector,
+    viz => viz.hierarchialData || emptyObject,
+);
+
+export const chordDataSelector = createSelector(
+    visualizationSelector,
+    viz => viz.chordData || emptyObject,
+);
+
+export const correlationDataSelector = createSelector(
+    visualizationSelector,
+    viz => viz.correlationData || emptyObject,
+);
+
+export const barDataSelector = createSelector(
+    visualizationSelector,
+    viz => viz.barData || emptyObject,
+);
+
+export const forceDirectedDataSelector = createSelector(
+    visualizationSelector,
+    viz => viz.forceDirectedData || emptyObject,
+);
+
+export const usersInformationListSelector = createSelector(
+    usersSelector,
+    users => (
+        Object.keys(users).map(
+            id => users[id].information,
+        ) || emptyList
+    ).filter(d => d),
+);
+
+export const regionsListSelector = createSelector(
+    regionsSelector,
+    regions => (
+        (regions && Object.values(regions).filter(
+            region => region && region.public,
+        )) || emptyList
+    ),
+);
+
+export const analysisFrameworkListSelector = createSelector(
+    analysisFrameworksSelector,
+    analysisFrameworks => (
+        (analysisFrameworks && Object.values(analysisFrameworks).filter(
+            analysisFramework => analysisFramework,
+        )) || emptyList
+    ),
+);
+
+export const categoryEditorListSelector = createSelector(
+    categoryEditorsSelector,
+    categoryEditors => (
+        (categoryEditors && Object.values(categoryEditors).filter(
+            categoryEditor => categoryEditor,
+        )) || emptyList
+    ),
+);
+
+// MODIFERS
+
+// activeUser
+const currentUserSelector = createSelector(
+    activeUserSelector,
+    usersSelector,
+    (activeUser, users) => (users[activeUser.userId] || emptyObject),
+);
+
+// userIdFromRoute
+const userSelector = createSelector(
+    userIdFromRoute,
+    usersSelector,
+    (userId, users) => (users[userId] || emptyObject),
+);
+
+// OTHERS
+
+// regionIdFromProps
+export const regionDetailForRegionSelector = createSelector(
+    regionsSelector,
+    regionIdFromProps,
+    (regions, regionId) => (regions[regionId] || emptyObject),
+);
+
+// regionIdFromProps
 export const adminLevelForRegionSelector = createSelector(
     adminLevelsSelector,
     regionIdFromProps,
@@ -104,120 +162,151 @@ export const adminLevelForRegionSelector = createSelector(
     ),
 );
 
-export const regionDetailForRegionSelector = createSelector(
-    regionsSelector,
-    regionIdFromProps,
-    (regions, regionId) => (
-        regions[regionId] || emptyObject
+// countryIdFromProps
+export const countryDetailSelector = createSelector(
+    regionsListSelector,
+    countryIdFromProps,
+    (regions, activeCountry) => (
+        regions.find(
+            country => country.id === +activeCountry,
+        ) || emptyObject
     ),
 );
 
-// Selector depending on user id from route (url)
-
-export const userSelector = createSelector(
-    userIdFromRoute,
-    usersSelector,
-    (userId, users) => (users[userId] || emptyObject),
+// analysisFrameworkIdFromPropsForProject
+export const analysisFrameworkDetailSelector = createSelector(
+    analysisFrameworksSelector,
+    analysisFrameworkIdFromPropsForProject,
+    (analysisFrameworks, afId) => (
+        analysisFrameworks[afId] || emptyObject
+    ),
 );
 
-export const userInformationSelector = createSelector(
-    userSelector,
-    user => (user.information || emptyObject),
-);
-export const usersInformationListSelector = createSelector(
-    usersSelector,
-    users => (Object.keys(users).map(id =>
-        users[id].information,
-    ) || emptyList).filter(d => d),
+// categoryEditorIdFromPropsForProject
+export const categoryEditorDetailSelector = createSelector(
+    categoryEditorsSelector,
+    categoryEditorIdFromPropsForProject,
+    (categoryEditors, ceId) => (
+        categoryEditors[ceId] || emptyObject
+    ),
 );
 
-export const userProjectsSelector = createSelector(
-    projectsSelector,
-    userSelector,
-    (projects, user) => ((user.projects &&
-        user.projects.map(projectId => (
-            projects[projectId]
-        ))) || emptyList),
-);
-
-export const userGroupsSelector = createSelector(
-    groupsSelector,
-    userSelector,
-    (userGroups, user) => ((user.userGroups &&
-        user.userGroups.map(userGroupId => (
-            userGroups[userGroupId]
-        ))) || emptyList),
-);
-
-export const userGroupProjectSelector = createSelector(
-    projectsSelector,
-    groupIdFromRoute,
-    (projects, userGroupId) => (Object.keys(projects).reduce((acc, projectId) => {
-        const userGroups = (projects[projectId] || emptyObject).userGroups;
-        if (userGroups && userGroups.find(userGroup => (userGroup.id === +userGroupId))) {
-            acc.push(projects[projectId]);
-        }
-        return acc;
-    }, []) || emptyList),
-);
-
+// groupIdFromRoute
 export const groupSelector = createSelector(
     groupsSelector,
     groupIdFromRoute,
     (userGroups, userGroupId) => (userGroups[userGroupId] || emptyObject),
 );
 
+// groupIdFromRoute
+export const userGroupProjectSelector = createSelector(
+    projectsSelector,
+    groupIdFromRoute,
+    (projects, userGroupId) => (
+        Object.keys(projects)
+            .reduce(
+                (acc, projectId) => {
+                    const userGroups = (projects[projectId] || emptyObject).userGroups;
+                    const hasUserGroup = userGroups && userGroups.find(
+                        userGroup => (userGroup.id === +userGroupId),
+                    );
+                    if (hasUserGroup) {
+                        return [
+                            ...acc,
+                            projects[projectId],
+                        ];
+                    }
+                    return acc;
+                },
+                emptyList,
+            )
+    ),
+);
+
+// userGroupIdFromProps
 export const userGroupDetailsSelector = createSelector(
     groupsSelector,
     userGroupIdFromProps,
     (userGroups, userGroupId) => (userGroups[userGroupId] || emptyObject),
 );
-// Selector depending on user id from state (logged-in user)
 
-export const currentUserSelector = createSelector(
-    activeUserSelector,
-    usersSelector,
-    (activeUser, users) => (users[activeUser.userId] || emptyObject),
+// userIdFromRoute
+export const userInformationSelector = createSelector(
+    userSelector,
+    user => (user.information || emptyObject),
 );
 
+// userIdFromRoute
+export const userProjectsSelector = createSelector(
+    projectsSelector,
+    userSelector,
+    (projects, user) => (
+        (user.projects && user.projects.map(
+            projectId => projects[projectId],
+        )) || emptyList
+    ),
+);
+
+// userIdFromRoute
+export const userGroupsSelector = createSelector(
+    groupsSelector,
+    userSelector,
+    (userGroups, user) => (
+        (user.userGroups && user.userGroups.map(
+            userGroupId => (userGroups[userGroupId]),
+        )) || emptyList
+    ),
+);
+
+// activeUser
 export const currentUserInformationSelector = createSelector(
     currentUserSelector,
     user => (user.information || emptyObject),
 );
 
+// activeUser
 export const currentUserProjectsSelector = createSelector(
     currentUserSelector,
     projectsSelector,
-    (user, projects) => ((user.projects &&
-        user.projects.map(projectId => (
-            projects[projectId] || emptyObject
-        ))
-    ) || emptyList),
+    (user, projects) => (
+        (user.projects && user.projects.map(projectId =>
+            projects[projectId] || emptyObject,
+        )) || emptyList
+    ),
 );
 
+// activeUser
 export const currentUserAdminProjectsSelector = createSelector(
     currentUserProjectsSelector,
     projects => projects.filter(project => (project.role === 'admin')),
 );
 
-// Visualization Selectors
-
-export const hierarchialDataSelector = ({ domainData }) => (
-    domainData.visualization.hierarchialData || emptyObject
+// activeUser, projectIdFromRoute
+export const currentUserActiveProjectSelector = createSelector(
+    currentUserProjectsSelector,
+    projectIdFromRoute,
+    (currentUserProjects, activeProject) => (
+        currentUserProjects.find(project => project.id === activeProject) || emptyObject
+    ),
 );
 
-export const chordDataSelector = ({ domainData }) => (
-    domainData.visualization.chordData || emptyObject
+// projectIdFromRoute
+export const leadFilterOptionsForProjectSelector = createSelector(
+    projectIdFromRoute,
+    leadFilterOptionsSelector,
+    (activeProject, leadFilterOptions) => (leadFilterOptions[activeProject] || emptyObject),
 );
 
-export const correlationDataSelector = ({ domainData }) => (
-    domainData.visualization.correlationData || emptyObject
+// projectIdFromRoute
+export const projectDetailsSelector = createSelector(
+    projectsSelector,
+    projectIdFromRoute,
+    (projects, activeProject) => projects[activeProject] || emptyObject,
 );
 
-export const barDataSelector = ({ domainData }) => (
-    domainData.visualization.barData || emptyObject
-);
-
-export const forceDirectedDataSelector = ({ domainData }) => (
-    domainData.visualization.forceDirectedData || emptyObject
+// projectIdFromRoute
+export const projectOptionsSelector = createSelector(
+    projectsOptionsSelector,
+    projectIdFromRoute,
+    (projectsOptions, activeProject) => projectsOptions[activeProject] || emptyObject,
 );
