+++ frontend/src/common/selectors/siloDomainData/common.js
@@ -1,0 +1,9 @@
+// TODO: save match directly and get data from there
+
+export const activeProjectSelector = ({ siloDomainData }) => (
+    siloDomainData.activeProject
+);
+
+export const activeCountrySelector = ({ siloDomainData }) => (
+    siloDomainData.activeCountry
+);
