+++ frontend/src/common/selectors/siloDomainData/galleryFiles.js
@@ -1,0 +1,7 @@
+const emptyList = [];
+// Gallery Files
+
+// eslint-disable-next-line import/prefer-default-export
+export const userGalleryFilesSelector = ({ siloDomainData }) => (
+    siloDomainData.userGalleryFiles || emptyList
+);
