+++ frontend/src/topic/Project/components/ProjectAnalysisFramework/index.js
@@ -218,6 +218,7 @@ export default class ProjectAnalysisFramework extends React.PureComponent {
 
         return (
             <ProjectAfDetail
+                match={this.props.match}
                 mainHistory={this.props.mainHistory}
                 key={selectedAf}
                 afId={selectedAf}
