+++ frontend/src/topic/Project/components/ProjectCategoryEditor/index.js
@@ -218,6 +218,7 @@ export default class ProjectCategoryEditor extends React.PureComponent {
 
         return (
             <ProjectCeDetail
+                match={this.props.match}
                 mainHistory={this.props.mainHistory}
                 key={selectedCe}
                 ceId={selectedCe}
