+++ frontend/src/topic/Project/components/ProjectRegions/index.js
@@ -162,6 +162,7 @@ export default class ProjectRegions extends React.PureComponent {
         if ((projectDetails.regions || emptyList).length > 0) {
             return (
                 <ProjectRegionDetail
+                    match={this.props.match}
                     key={selectedRegion}
                     regionId={selectedRegion}
                 />
@@ -236,6 +237,7 @@ export default class ProjectRegions extends React.PureComponent {
                                 }
                                 {selectedAddRegionOption === 'old' &&
                                     <AddExistingRegion
+                                        match={this.props.match}
                                         styleName="add-existing-region"
                                         projectId={projectDetails.id}
                                         onModalClose={this.handleModalClose}
